# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "用户登录获取访问令牌",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  }
                },
                "required": [
                  "username",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功"
          },
          "401": {
            "description": "用户名或密码错误"
          }
        },
        "summary": "用户登录",
        "tags": [
          "认证"
        ]
      }
    },
    "/auth/send-code": {
      "post": {
        "description": "向指定邮箱发送验证码",
        "operationId": "AuthController_sendVerificationCode",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "description": "邮箱地址"
                  }
                },
                "required": [
                  "email"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "验证码发送成功"
          },
          "409": {
            "description": "邮箱已被注册"
          }
        },
        "summary": "发送验证码",
        "tags": [
          "认证"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "用户注册并自动登录",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "用户名（唯一）"
                  },
                  "email": {
                    "type": "string",
                    "description": "邮箱（唯一）"
                  },
                  "password": {
                    "type": "string",
                    "description": "密码"
                  },
                  "code": {
                    "type": "string",
                    "description": "验证码"
                  },
                  "bio": {
                    "type": "string",
                    "description": "个人简介（可选）"
                  },
                  "nickname": {
                    "type": "string",
                    "description": "用户昵称（可选）"
                  }
                },
                "required": [
                  "username",
                  "email",
                  "password",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "注册成功"
          },
          "400": {
            "description": "验证码错误或已过期"
          },
          "409": {
            "description": "用户名或邮箱已存在"
          }
        },
        "summary": "用户注册",
        "tags": [
          "认证"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "使用现有令牌刷新获取新令牌",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "刷新令牌"
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "令牌刷新成功"
          },
          "401": {
            "description": "令牌无效"
          }
        },
        "summary": "刷新令牌",
        "tags": [
          "认证"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "description": "获取当前登录用户的信息",
        "operationId": "AuthController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取用户信息",
        "tags": [
          "认证"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "用户注销登录",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "注销成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "用户注销",
        "tags": [
          "认证"
        ]
      }
    },
...
    "/focus-records": {
      "post": {
        "description": "创建一条新的专注记录",
        "operationId": "FocusRecordController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "专注记录创建数据",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "创建专注记录",
        "tags": [
          "专注记录"
        ]
      },
      "get": {
        "description": "获取当前用户的所有专注记录列表",
        "operationId": "FocusRecordController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取所有专注记录",
        "tags": [
          "专注记录"
        ]
      }
    },
    "/focus-records/{id}": {
      "get": {
        "description": "根据ID获取特定的专注记录详情",
        "operationId": "FocusRecordController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "专注记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "记录不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取单个专注记录",
        "tags": [
          "专注记录"
        ]
      },
      "put": {
        "description": "更新指定ID的专注记录信息",
        "operationId": "FocusRecordController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "专注记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "专注记录更新数据",
          "content": {
            "application/json": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "记录不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "更新专注记录",
        "tags": [
          "专注记录"
        ]
      },
      "delete": {
        "description": "删除指定ID的专注记录",
        "operationId": "FocusRecordController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "专注记录ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "401": {
            "description": "未授权"
          },
          "404": {
            "description": "记录不存在"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "删除专注记录",
        "tags": [
          "专注记录"
        ]
      }
    },
    "/focus-records/task/{taskId}": {
      "get": {
        "description": "获取指定任务的所有专注记录",
        "operationId": "FocusRecordController_findByTaskId",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "description": "任务ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取任务专注记录",
        "tags": [
          "专注记录"
        ]
      }
    },
    "/focus-records/stats/summary": {
      "get": {
        "description": "获取当前用户的专注统计信息",
        "operationId": "FocusRecordController_getStatistics",
        "parameters": [],
        "responses": {
          "200": {
            "description": "获取成功"
          },
          "401": {
            "description": "未授权"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "获取专注统计",
        "tags": [
          "专注记录"
        ]
      }
    },
...
  },
  "info": {
    "title": "叮咚待办 (DingDongTodo) API",
    "description": "DingDongTodo后端API文档",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT认证Bearer Token",
        "in": "header"
      }
    },
...
  }
}